Move e2e test app in e2e/browser/webapp, don't reuse codesandbox
