+++ .gitignore
@@ -15,3 +15,4 @@ docs/dist/
 test-results/
 .vercel
 .virtualenv
+.parcel-cache/
