+++ e2e/browser/webapp/.gitignore
@@ -1,0 +1,3 @@
+.parcel-cache
+dist
+node_modules
