+++ e2e/browser/webapp/src/index.html
@@ -7,10 +7,10 @@
 
 <body>
 	<label for="idp">Identity provider:</label>
-	<input type="url" name="idp" id="idp" value="https://broker.pod.inrupt.com"/>
+	<input type="url" name="idp" id="idp" value="https://broker.pod.inrupt.com" />
 	<button>Log in</button>
 
-	<script src="src/end-to-end-test.ts" type="module">
+	<script src="index.ts" type="module">
 	</script>
 </body>
